Fixed batchsize param type

+++ ch-covidcertificate-backend-delivery/ch-covidcertificate-backend-delivery-ws/src/main/java/ch/admin/bag/covidcertificate/backend/delivery/ws/config/WsBaseConfig.java
@@ -70,7 +70,7 @@ public abstract class WsBaseConfig implements WebMvcConfigurer {
     private String iosPushTopic;
 
     @Value("${push.batchsize}")
-    private long batchSize;
+    private int batchSize;
 
     public abstract DataSource dataSource();
